* Support version 1 mdhd
git-svn-id: https://svn.code.sf.net/p/gmerlin/code/trunk/gmerlin_avdecoder@4103 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ include/qt.h
@@ -840,10 +840,10 @@ typedef struct
 
   int version;
   uint32_t flags;
-  uint32_t creation_time;
-  uint32_t modification_time;
+  uint64_t creation_time;
+  uint64_t modification_time;
   uint32_t time_scale;
-  uint32_t duration;
+  uint64_t duration;
   uint16_t language;
   uint16_t quality;
   } qt_mdhd_t;
